requirements.txt
angr
unicorn
qiling
r2pipe
pwntools
lief
